Started on RenderCallbackBuffer abstraction
This is analogous to CPAL's UnknownTypeBuffer, and is used basically as
an admission of defeat in terms of determining the type of the samples
for an AudioUnit's rendering callback at compile time, which is
impossible since the user can change the stream type dynamically. Using
this construct, we can provide a safe abstraction over the data in the
buffers.